InputManual component now working.
